refactor: remove obsolete rules forwarding module
