+++ packages/zodvex/__tests__/rules.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { z } from 'zod'
-import { ZodvexDatabaseReader, ZodvexDatabaseWriter } from '../src/internal/db'
+import { RulesQueryChain, ZodvexDatabaseReader, ZodvexDatabaseWriter } from '../src/internal/db'
 import type {
   ReaderAuditConfig,
   TableRules,
@@ -8,8 +8,7 @@ import type {
   WriterAuditConfig,
   ZodvexRules,
   ZodvexRulesConfig
-} from '../src/internal/rules'
-import { RulesQueryChain } from '../src/internal/rules'
+} from '../src/internal/ruleTypes'
 import type { ZodTableSchemas } from '../src/internal/schema'
 import { zx } from '../src/internal/zx'
