--- packages/zodvex/src/internal/rules.ts
@@ -1,19 +1,0 @@
-// The DB module owns the base classes and their rule/audit subclasses so
-// constructors are ready synchronously, without a cross-module installer.
-export { normalizeReadResult, RulesQueryChain } from './db'
-
-export {
-  type DeleteRule,
-  type InsertDoc,
-  type InsertRule,
-  type PatchRule,
-  type ReaderAuditConfig,
-  type ReadRule,
-  type ReplaceRule,
-  type ResolveDecodedDocForRules,
-  type TableRules,
-  type WriteEvent,
-  type WriterAuditConfig,
-  type ZodvexRules,
-  type ZodvexRulesConfig
-} from './ruleTypes'
